requirements.txt
arrow==0.15.2
astroid==2.3.3
autopep8==1.5
Babel==2.7.0
beautifulsoup4==4.8.0
bleach==3.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
Click==7.0
cssmin==0.2.0
cssprefixer==1.4.1
cssselect2==0.3.0
cssutils==1.0.2
defusedxml==0.6.0
Django==2.2.2
django-appconf==1.0.3
django-autocomplete-light==3.4.1
django-chartjs==2.1.0
django-compressor==2.2
django-contrib-comments==1.9.1
django-crispy-forms==1.7.2
django-environ==0.4.5
django-extensions==2.2.1
-e git+https://github.com/JordanReiter/django-impersonate-auth.git@5f91ca250460bcbf8638c42aef140999b04a0a90#egg=django_impersonate_auth
django-navutils==0.7
django-phonenumber-field==3.0.1
django-private-storage==2.2
django-reversion==3.0.8
django-session-security==2.6.6
django-simple-pagination==1.3
django-weasyprint==0.5.5
django-widget-tweaks==1.4.5
djangorestframework==3.11.1
docutils==0.15.2
easy-thumbnails==2.6
Flask==1.1.1
Flask-Assets==0.12
gevent==1.4.0
gevent-socketio-hartwork==0.3.6.post201803091952
gevent-websocket==0.10.1
greenlet==0.4.15
html5lib==1.0.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
Logbook==1.5.2
mailtrap==0.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
passlib==1.7.1
persisting-theory==0.2.1
phonenumbers==8.10.19
Pillow==6.0.0
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
Pyphen==0.9.5
pyScss==1.3.5
python-daemon==2.2.3
python-dateutil==2.8.0
python-http-client==3.1.0
pytz==2019.1
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.0.12
sendgrid==6.0.5
six==1.12.0
sorl-thumbnail==12.5.0
soupsieve==1.9.3
sqlparse==0.3.0
tinycss2==1.0.2
typed-ast==1.4.1
urllib3==1.25.3
WeasyPrint==51
webassets==0.12.1
webencodings==0.5.1
Werkzeug==0.15.6
wrapt==1.11.2
XlsxWriter==1.2.1
